+++ host/examples/rfnoc_rx_to_file.cpp
@@ -188,7 +188,7 @@ bool check_locked_sensor(std::vector<std::string> sensor_names, const char* sens
 
     while (true) {
         if ((not first_lock_time.is_not_a_date_time()) and
-                (boost::get_system_time() > (first_lock_time + boost::posix_time::seconds(setup_time))))
+                (boost::get_system_time() > (first_lock_time + boost::posix_time::seconds(static_cast<long>(setup_time)))))
         {
             std::cout << " locked." << std::endl;
             break;
@@ -202,7 +202,7 @@ bool check_locked_sensor(std::vector<std::string> sensor_names, const char* sens
         else {
             first_lock_time = boost::system_time();	//reset to 'not a date time'
 
-            if (boost::get_system_time() > (start + boost::posix_time::seconds(setup_time))){
+            if (boost::get_system_time() > (start + boost::posix_time::seconds(static_cast<long>(setup_time)))){
                 std::cout << std::endl;
                 throw std::runtime_error(str(boost::format("timed out waiting for consecutive locks on sensor \"%s\"") % sensor_name));
             }
